+++ CHANGELOG.md
@@ -2,12 +2,25 @@
 All notable changes to this project will be documented in this file.
 
 #### 1.x Releases
-- `1.6.x` Releases  - [1.6.0](#160)
+- `1.6.x` Releases  - [1.6.0](#160) | [1.6.1](#161) | [1.6.2](#162)
 - `1.5.x` Releases  - [1.5.0](#150)
 - `1.4.x` Releases  - [1.4.0](#140)
 - `1.3.x` Releases  - [1.3.0](#130)
 - `1.2.x` Releases  - [1.2.0](#120) | [1.2.3](#123)  | [1.2.4](#124)
 - `1.1.x` Releases  - [1.1.0](#110) | [1.1.1](#111) | [1.1.2](#112) | [1.1.3](#113) | [1.1.4](#114)  | [1.1.5](#115) | [1.1.6](#116) | [1.1.7](#117) [1.1.8](#118) [1.1.9](#119) 
+
+---
+## [1.6.2](https://github.com/congncif/IDMFoundation/releases/tag/1.6.2)
+Released on 2018-6-6
+
+- Added more custom methods for data provider
+
+---
+## [1.6.1](https://github.com/congncif/IDMFoundation/releases/tag/1.6.1)
+Released on 2018-6-6
+
+- Removed extra data from response protocol
+
 ---
 ## [1.6.0](https://github.com/congncif/IDMFoundation/releases/tag/1.6.0)
 Released on 2018-5-30